Subject: Renaming of generic_reacher
Detailed Info:
Answers comments by @c-salmi
To avoid confusion with the abstract class GenericRobot, the generic
urdf reacher is renamed: GenericRobot was moved to GenericUrdfReacher. The
environments and the folder were moved to generic_urdf_reacher.
Changes the example for the ur5 robot from torque interface to velocity
interface to have a nice first example.

Issue: #104