ft: Moves rendering of goals and obstacles inside this package.
Rendering and adding of obstacles and goals used to be done inside
mpscenes. With the aim to be engine agnostic, the rendering are moved
inside this package.

Interesting to @c-salmi for the fix on collision margins. The same
issues occurs now with sphere as they no longer use a mesh but rather
the default pybullet types.