ft[structure]: Redefines reset and __init__in urdf_env.
The goal is to make sure that the reset function can be used without
re-adding obstacles, goals and sensors. The obstacles are now set before
the reset function is called. That requires to set the spaces
explicitely after all obstacles, goals and sensors have been added.
Then, this allows to use the sensors with the FlattenObservation
wrapper. This allowed to remove the option `flatten_observation` all
together.