ft[gymnasium]: Updates step and reset api.
Gymnasium encourages the change of the return values of the reset and
step functions. Now, the reset-function returns the initial observation
and an info dictionary (in our case still an empty one). The step
function was changed to return not only the done boolean but two
boolean, one for terminated and one for truncated. So far, our
implementation has no truncation criteria, so it is always false.